created a separate module for the main group of endpoints
